Move pyrsistent serialization and hashing helpers to a new module. Re-organize the tests accordingly.
